// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/warehouse/models/warehouse.model.ts
export class WarehouseModel {
  WarehouseId: string;
  WarehouseCode: string;
  WarehouseName: string;
  WarehouseParent: string;
  WarehouseAddress: string;
  WarehousePhone: string;
  Storagekeeper: string;
  WarehouseDescription: string;
  Active: boolean;
  CreatedDate: Date;
  CreatedById: string;
  UpdatedDate: Date;
  UpdatedById: string;
}